Implement Car reservation from navigation panel
