+++ src/components/cars/DeleteCar.js
@@ -1,12 +1,13 @@
 import React from 'react';
+import CarCard from './CarCard';
 // /* eslint-disable react/prop-types */
 // import React from 'react';
 // import React from 'react';
-import Cars from './Cars';
+// import Cars from './Cars';
 
 const DeleteCar = () => (
   <div>
-    <Cars showButton />
+    <CarCard showButton />
   </div>
 );
