+++ src/components/cars/ReserveHome.js
@@ -2,26 +2,15 @@ import React from 'react';
 // import { useSelector, useDispatch } from 'react-redux';
 // import { fetchUser } from '../../redux/features/user';
 // import { useLocation } from 'react-router-dom';
-import Cars from './Cars';
+// import Cars from './Cars';
+import CarCard from './CarCard';
 
-const ReserveHome = () => {
-  // const dispatch = useDispatch();
-  // const getAllCars = useSelector((state) => state.cars);
-  // const user = useSelector((state) => state.user);
-  console.log('user');
-  // useEffect(() => {
-  //   dispatch(fetchUser());
-  // }, [dispatch]);
-  // console.log(getAllCars);
-
-  return (
+const ReserveHome = () => (
+  (
     <div>
-      <div>
-        <h2>Reserve a car</h2>
-      </div>
-      <Cars showHeader={false} />
+      <CarCard showHeader />
     </div>
-  );
-};
+  )
+);
 
 export default ReserveHome;
